requirements.txt
segno
Pillow